Update cardIndicators field to fix bug
Confirm that card_indicators is being passed to the gateway. We have 1 test failing with this message: @message="Request does not adhere to the DTD. Please correct and send again.", but I was able to confirm that card_indicators is being passed with options. See options hash below. 3DS tests on orbital are still failing, as they have been on master.

{:order_id=>"bc8b1fd1ca40f4d8ba6e072cfa11f38f", :address=>{:name=>"Jim Smith", :address1=>"456 My Street", :address2=>"Apt 1", :company=>"Widgets Inc", :city=>"Ottawa", :state=>"ON", :zip=>"K1C2N6", :country=>"CA", :phone=>"(555)555-5555", :fax=>"(555)555-6666"}, :merchant_id=>"merchant1234", :soft_descriptors=>{:merchant_name=>"Merch", :product_description=>"Description", :merchant_email=>"email@example"}, :card_indicators=>"y"}

Unit:
95 tests, 561 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Remote:
36 tests, 184 assertions, 5 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
86.1111% passed